requirements.txt
Flask==3.0.0
Pillow>=10.0.0
reportlab>=4.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
pdf2docx>=0.5.8
